✅ 已修复！speed = location.speed.toDouble() 改为 speed = location.speed。

+++ app/src/main/java/com/runshare/app/ui/screens/HomeScreen.kt
@@ -461,7 +461,7 @@ private fun getIdleLocation(
                         latitude = location.latitude,
                         longitude = location.longitude,
                         altitude = location.altitude,
-                        speed = location.speed.toDouble(),
+                        speed = location.speed,
                         timestamp = System.currentTimeMillis()
                     )
                 )
